Add authentication check before controller actions;

+++ app/Http/Controllers/Dashboard/ProductController.php
@@ -3,11 +3,11 @@
 namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests;
 use App\Http\Requests\ProductUpdateRequest;
 use App\Product;
 use Illuminate\Http\Request;
-
-use App\Http\Requests;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
@@ -25,12 +25,15 @@ class ProductController extends Controller
     public function edit($productId)
     {
         $product = Product::find($productId);
+        if (Gate::denies('product', $product)) abort(403);
+
         return view('dashboard.product.edit', ['product' => $product]);
     }
 
     public function update($productId, ProductUpdateRequest $request)
     {
         $product = Product::find($productId);
+        if (Gate::denies('product', $product)) abort(403);
 
         $fields = ['lm', 'name', 'category', 'free_shipping', 'description', 'price'];
         foreach ($fields as $field) {
@@ -45,6 +48,8 @@ class ProductController extends Controller
     public function destroy($productId)
     {
         $product = Product::find($productId);
+        if (Gate::denies('product', $product)) abort(403);
+
         $product->delete();
 
         return redirect()->route('product.index')->with('message', 'Product deleted');
